requirements.txt
streamlit>=1.13.0
pillow>=9.0.0
gtts>=2.3.0
numpy>=1.22.0